Auth by key maintenance

+++ temboardui/plugins/maintenance/static/js/temboard.maintenance.schema.js
@@ -58,14 +58,14 @@ $(function() {
       sortBy: sortBy,
       indexSortBy: indexSortBy,
       doReindex: doReindex,
-      cancelReindex: cancelReindex
+      cancelReindex: cancelReindex,
+      checkSession: checkSession
     }
   });
 
   function getSchemaData() {
     $.ajax({
       url: apiUrl,
-      beforeSend: function(xhr){xhr.setRequestHeader('X-Session', xsession);},
       contentType: "application/json",
       success: (function(data) {
         this.schema = data;
@@ -165,6 +165,9 @@ $(function() {
   }
 
   function cancelReindex(id) {
+    if (!checkSession()) {
+      return;
+    }
     $.ajax({
       method: 'DELETE',
       url: maintenanceBaseUrl + '/reindex/' + id,
@@ -184,5 +187,25 @@ $(function() {
   function indexSortBy(criteria, order) {
     this.indexSortCriteria = criteria;
     this.indexSortOrder = order || 'asc';
+  }
+
+  /**
+   * Redirects to agent login page if session is not provided
+   * Should be used in each action requiring xsession authentication.
+   *
+   * params:
+   * e - Optional browser event
+   *
+   */
+  function checkSession(e) {
+    if (!xsession) {
+      var params = $.param({redirect_to: window.location.href});
+      if (confirm('You need to be logged in the instance agent to perform this action')) {
+        window.location.href = agentLoginUrl + '?' + params;
+      }
+      e && e.stopPropagation();
+      return false;
+    }
+    return true;
   }
 });
